controller

+++ src/Constants.java
@@ -3,9 +3,9 @@ import java.awt.*;
 public class Constants {
     public static final int SCREEN_WIDTH = 800;
     public static final int SCREEN_HEIGHT = 600;
-    public static final int OBJECT_WIDTH = 10;
-    public static final int OBJECT_HEIGHT = 100;
+    public static final double OBJECT_WIDTH = 10;
+    public static final double OBJECT_HEIGHT = 100;
     public static final Color OBJECT_COLOR = Color.WHITE;
-    public static final int BALL_WIDTH = 10;
-    public static final int HORIZONTAL_PADDING = 40;
+    public static final double BALL_WIDTH = 10;
+    public static final double HORIZONTAL_PADDING = 40;
 }
